requirements.txt
st-files-connection
s3fs
langchain>=0.0.219
boto3<=1.33.1
openai
anthropic
streamlit
python-dotenv
